Per-Part Calculation for Blast Resistance
After fixing the armor bug, it was noted that current blast mechanics would significantly under-estimate blast penetration. Despite being able to, according to the equations we use, blow through the armor, the blast is basically completely stopped a couple of m past the armor. Assessing blast pressure at the part, and then scaling overall blast pressure by this fixes this issue.

It must be noted that as it is right now, big explosions are consistently overpenetrating, which is something to look into and re-assess in terms of the equations.